+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+* Removed the unused `Model3dSceneObject`.
 
 ## [0.4.4] 2024-09-03